created the axios.post endpoint with authentication
